requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pandas==2.1.4
numpy==1.26.3
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2
yfinance==0.2.32
ta==0.10.2
python-decouple==3.8
pytz==2024.1
scikit-learn==1.3.2
tensorflow==2.14.0
keras==2.14.0
pandas-market-calendars==4.3.2